fixing some issues with the search functionality and restructuring code
