+++ create_post.html
@@ -41,8 +41,8 @@
         </div>
     </nav>
     </header>
-    <div class="create_ post flex-grow-2">
-        <div class="container-fluid">
+    <div class="create_ post flex-grow-2 mt-5">
+        <div class="d-flex justify-content-center">
             <div class="col-12 col-md-6">
                 <form id="create_post">
                     <div class="form-group mb-4">
