+++ js/posts/create.mjs
@@ -3,72 +3,35 @@ import { fetchWithToken } from "../posts/fetch_with_token.mjs";
 
 const method = "post";
 async function createPost(postData) {
-    
-    const response = await fetchWithToken(post, {
-        method, 
-        body: JSON.stringify(postData)
+  const response = await fetchWithToken(post, {
+    method,
+    body: JSON.stringify(postData),
+  });
 
-    });
+  const result = await response.json();
 
-    const result = await response.json();
-    
-   
-
-    console.log(result);
+  console.log(result);
 }
 
-
-
-
 function createFormListener() {
-    const form = document.querySelector("#create_post");
+  const form = document.querySelector("#create_post");
 
+  if (form) {
+    form.addEventListener("submit", (event) => {
+      event.preventDefault();
+      const form = event.target;
+      const formData = new FormData(form);
+      const post = Object.fromEntries(formData.entries());
 
-    if (form) {
-        form.addEventListener("submit", (event) => {
-            event.preventDefault();
-            const form = event.target;
-            const formData = new FormData(form);
-            const post = Object.fromEntries(formData.entries())
+      console.log(post);
 
-            console.log(post);
-            
-            
-    
-            createPost(post);
+      createPost(post);
 
-            /*setTimeout( () => {
+      setTimeout( () => {
                 window.open("index.html", "_self");
-            }, 1500);*/
-            
-        } );
-    
-    }  
-   
-
-   
-   
-    
+            }, 1500);
+    });
+  }
 }
 
 createFormListener();
-
-
-
-//delete id 906, 913, 915, 1238, 1239
-    
-
-
-
-   
- 
-  
-
-
-/*createPost({
-    title: "testPostAbra",
-    body: "deleted sooon",
-    media: "https://picsum.photos/200/200"
-
-});*/
-
