+++ js/posts/update.mjs
@@ -1,6 +1,5 @@
-import { post, post_author,  } from "../API_URLs_export.mjs";
+import { post, post_author } from "../API_URLs_export.mjs";
 import { fetchWithToken } from "../posts/fetch_with_token.mjs";
-
 
 const queryString = document.location.search;
 const params = new URLSearchParams(queryString);
@@ -8,52 +7,33 @@ const id = params.get("id");
 
 const postURL = post + id + post_author;
 
-
 const method = "put";
 async function updatePost(postData) {
-    
-    const response = await fetchWithToken(postURL, {
-        method, 
-        body: JSON.stringify(postData)
+  const response = await fetchWithToken(postURL, {
+    method,
+    body: JSON.stringify(postData),
+  });
 
-    });
-
-    const result = await response.json();
-   
-    
-    
-
-    
+  const result = await response.json();
 }
 
-
-
 function createFormListener() {
-    const form = document.querySelector("#update_post");
+  const form = document.querySelector("#update_post");
 
+  if (form) {
+    form.addEventListener("submit", (event) => {
+      event.preventDefault();
+      const form = event.target;
+      const formData = new FormData(form);
+      const post = Object.fromEntries(formData.entries());
 
-    if (form) {
-        form.addEventListener("submit", (event) => {
-            event.preventDefault();
-            const form = event.target;
-            const formData = new FormData(form);
-            const post = Object.fromEntries(formData.entries())
+      updatePost(post);
 
-            
-            updatePost(post);
-            
-            setTimeout( () => {
-                window.location.reload(true);
-            }, 1000);
-        
-        } );
-    
-    }  
-   
-
-   
-   
-    
+      setTimeout(() => {
+        window.location.reload(true);
+      }, 1000);
+    });
+  }
 }
 
 createFormListener();
