+++ post.html
@@ -45,7 +45,10 @@
         <h1 class="mb-3 fw-light"></h1>
         
     </div>
-    <button class="update_post_button">update</button>
+    <div class="button_update d-flex justify-content-center align-items-center ">
+        <button class="update_post_button mb-4">Update</button>
+    </div>
+    
 
     <div class="flex-grow-1" >
         <div class="d-flex justify-content-center">
@@ -67,7 +70,7 @@
                         <label for="">media</label>
                         <input type="url" class="form-control" id="media" name="media" required>
                       </div>
-                     <!-- <button type="button" class="create_post_button btn btn-primary btn-lg mt-3">create post</button>-->
+                     
                      <button class="btn btn-success">update post</button>
                      
                   </form>
@@ -77,7 +80,7 @@
     
     
     <script src="/node_modules/bootstrap/dist/js/bootstrap.bundle.min.js"></script>
-    <script type="module" src="js/post.mjs"></script>   
+    <script type="module" src="js/post_by_id.mjs"></script>   
     <script type="module" src="js/posts/update.mjs"></script>  
     <script type="module" src="js/posts/remove.mjs"></script>
 </body>
