+++ src/scss/custom/_global.scss
@@ -11,8 +11,13 @@ html,
 }
 
 .update_post_button {
-  border-radius: 15px;
+  border-radius: 25px;
   border: none;
+  font-size: 30px;
+  background-color: rgb(10, 45, 121);
+  color: #fff;
+  
+  
 
 }
